+++ R/get-dataverse-data.R
@@ -106,7 +106,7 @@ get_organization_metadata <- function(organization = NULL) {
 #'
 #' This function uses [purrr::walk] together with `get_organization_metadata`
 #' to download Dataverse metadata from several organizations. Data are then
-#' stored as csv files into data-raw folder.
+#' stored as csv files into dataverse_raw folder.
 #'
 #' @param log_threshold The (standard Apache logj4) log level used as a
 #'   threshold for the logging infrastructure. See [logger::log_levels] for more
@@ -121,16 +121,13 @@ get_organization_metadata <- function(organization = NULL) {
 #' }
 get_dataverse_metadata <- function(log_threshold = logger::DEBUG) {
   logger::log_threshold(log_threshold)
-
-  logger::log_info("Cleaning old files")
-  pk_path <- system.file("data-raw", package = "aquadata.data.mapping")
-  csv_files <- list.files(pk_path, full.names = TRUE)
-  metadata_files <- csv_files[c(which(grepl("dataset_metadata", csv_files)))]
-  unlink(metadata_files, force = TRUE, recursive = TRUE)
-
   pars <- read_config()
 
-  logger::log_info("Downloading all Dataverse metadata")
+  logger::log_info("Cleaning old metadata")
+  folder_path <- system.file("dataverse_raw", package = "aquadata.data.mapping")
+  files <- list.files(folder_path, full.names = TRUE, recursive = TRUE)
+
+  logger::log_info("Downloading Dataverse raw metadata")
   pars$dataverse$organizations %>%
-  purrr::walk(get_organization_metadata)
+    purrr::walk(get_organization_metadata)
 }