+++ R/worldfish_guestbook_responses.R
@@ -1,7 +1,0 @@
-#' Worldfish Guest responses
-#'
-#' A dataset containing user activity responses on WordFish data from the Dataverse
-#' portal.
-#'
-#' @format A dataframe
-"worldfish_guestbook_responses"